Added advanced mode Layout
- Added terminal output for measurements
